Move code back from AnalysisPageViewModelState to AmalysisPane class
